System Prompt Crafter/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
lollms-client
ascii_colors
